+++ packages/react-form/src/index.ts
@@ -23,4 +23,5 @@ export {
   validateAll,
   reduceFields,
   makeCleanFields,
+  makeCleanDynamicList,
 } from './utilities';